Conflicts:
	Eurodroo/src/main/java/fr/esiea/glpoo/Eurodroo/Launcher.java

+++ Eurodroo/src/main/java/fr/esiea/glpoo/Eurodroo/domaine/dessin.java
@@ -4,17 +4,17 @@ import java.util.List;
 
 public interface dessin {
 
-	public double getAbscisse();
+	public int getAbscisse();
 	
-	public double getOrdonnée();
+	public int getOrdonnée();
 	
-	public double getTaille();
+	public int getTaille();
 	
 	public FormeDessin getForme();
 	
 	public List getCouleur();
 	
-	public double getRotation();
+	public int getRotation();
